+++ src/macros.rs
@@ -4,7 +4,7 @@ use util::Numeric;
 
 use std::collections::BTreeMap;
 
-pub struct FloatMap<T>(pub BTreeMap<OrderedFloat<f64>, T>);
+pub struct FloatMap<T>(pub BTreeMap<OrderedFloat<Numeric>, T>);
 impl<T> FloatMap<T> {
     pub fn default() -> Self {
         FloatMap(BTreeMap::new())
